ajout de su scan vidéo sur sanipasse borne
